Continued refactoring.
core
- Intrinsics learned fromMatrix().
- ArrayUtils learned to use template type deduction.
- ColorMap learned some linearRemapToLuminance().
- BinaryFileInputStream and BinaryFileOutputStream are now RAII objects.
- Minor refactoring of OBJ classes.
- Cleaned up MathUtils::rescale() interface.
- Documentation cleanup.

CUDA
- Added float3x3 and associated functions.

Kinect1x
- Fully removed dependency on Qt.
- KinectCamera is now an RAII class with a better polling interface.
- KinectCamera now supports infrared over the color channel.
- Added an experimental RGBDStream class for serializing data.

OpenCVInterop
- Removed nested namespaces.

+++ core/src/geometry/OpenNaturalCubicSpline.cpp
@@ -79,7 +79,7 @@ float OpenNaturalCubicSpline::evaluateAt( float t ) const
     int n;
     if( ct == 1.0f )
     {
-        n = m_vCoefficients.size() - 1;
+        n = static_cast< int >( m_vCoefficients.size() ) - 1;
     }
     else
     {
@@ -110,7 +110,7 @@ float OpenNaturalCubicSpline::derivativeAt( float t ) const
     int n;
     if( ct == 1.0f )
     {
-        n = m_vCoefficients.size() - 1;
+        n = static_cast< int >( m_vCoefficients.size() ) - 1;
     }
     else
     {
@@ -157,7 +157,7 @@ float OpenNaturalCubicSpline::inverse( float x, float tGuess, float epsilon,
 
 void OpenNaturalCubicSpline::computeCoefficients()
 {
-    int nPoints = m_vControlPoints.size();
+    int nPoints = static_cast< int >( m_vControlPoints.size() );
 
     std::vector< float > gamma( nPoints );
     std::vector< float > delta( nPoints );
